+++ z_tools.py
@@ -218,8 +218,6 @@ def CheckDeviceList(self, key, val) :
     '''
         This function is call during DeviceList load
     '''
-    import random
-
     Domoticz.Debug("CheckDeviceList - Address search : " + str(key))
     Domoticz.Debug("CheckDeviceList - with value : " + str(val))
 
@@ -294,11 +292,7 @@ def CheckDeviceList(self, key, val) :
             self.ListOfDevices[key]['ReadAttributes']=DeviceListVal['ReadAttributes']
         if 'IAS' in DeviceListVal :
             self.ListOfDevices[key]['IAS']=DeviceListVal['IAS']
-
-        # We will initialize Hearbeat with a random value between 0 to 12 in order to distribute the load when triggering action based on the Hearbeat value
-        # 12 is equivalent to 12 Heartbeat cycle ==> 2 minutes
-        self.ListOfDevices[key]['Heartbeat']=random.randint(0, 12)
-
+        self.ListOfDevices[key]['Heartbeat']=DeviceListVal['Heartbeat']
 
 def timeStamped( self, key, Type ):
     if key in self.ListOfDevices: