+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 - [Technical] Configurable plugin folders ( Data, Zdata, reports, www/templates )
 - [Technical] Confirgure by default the ZigBee Channels (recommended for Europe
 - [Hardware] Improve 2018 Xiaomi wireless sitches with 3 states per button ( lumi.remote.b286acn01, lumi.remote.b186acn01) 
+- [Technical] Remove the use of random
 
 ## 02 Decembre 2018 - Version 4.0.0
 - [Technical] New Zigate Transport layer  ( ZigBee compliant with retransmission in case of missing data, bit also a agressive algorithm.)